requirements.txt
Flask
requests
kafka-python
redis
PyMySQL
SQLAlchemy
scikit-learn
pandas
numpy
pyspark==3.5.0
confluent-kafka
pymysql
cryptography
python-dotenv